refactor: move NameChangeModal component
